python/requirements.txt
# Core
openai>=1.30,<2
qdrant-client>=1.9,<2
python-dotenv>=1.0,<2
pypdf>=4,<5
tiktoken>=0.7,<1
numpy>=1.26,<3

# Optional: Dev-Tools
# ruff>=0.4,<1
# black>=24,<25
